Re-organize client into seperate file
